support for 64-bit int fields

+++ Structorian.Engine/Fields/BitsField.cs
@@ -15,11 +15,11 @@ namespace Structorian.Engine.Fields
         {
             int offset = (int) reader.BaseStream.Position;
             IConvertible value = ReadIntValue(reader, instance);
-            uint u = value.ToUInt32(CultureInfo.CurrentCulture);
+            ulong u = value.ToUInt64(CultureInfo.CurrentCulture);
             StringBuilder displayValue = new StringBuilder(_size * 8);
             for(int i=_size * 8 - 1; i >= 0; i--)
             {
-                if ((u & (1 << i)) != 0)
+                if ((u & (1ul << i)) != 0)
                     displayValue.Append("1");
                 else
                     displayValue.Append("0");
