+++ Structorian.Engine/Fields/SetField.cs
@@ -15,20 +15,20 @@ namespace Structorian.Engine.Fields
         {
             int offset = (int)reader.BaseStream.Position;
             IConvertible iValue = ReadIntValue(reader, instance);
-            uint value = iValue.ToUInt32(CultureInfo.CurrentCulture);
+            ulong value = iValue.ToUInt64(CultureInfo.CurrentCulture);
             EnumDef enumDef = GetEnumAttribute("enum");
 
             StringBuilder result = new StringBuilder();
             for(int i=0; i<_size*8; i++)
             {
-                if ((value & (1 << i)) != 0)
+                if ((value & (1ul << i)) != 0)
                 {
                     if (result.Length > 0)
                         result.Append(", ");
                     result.Append(enumDef.ValueToString((uint) i));
                 }
             }
-            AddCell(instance, new EnumValue(value, result.ToString()), offset);
+            AddCell(instance, new EnumValue((uint) value, result.ToString()), offset);
         }
     }
 }
